refactor(schematics): divide functions
