+++ package/schematics/parse/parse/utils.angular.ts
@@ -1,6 +1,10 @@
 import { Tree } from '@angular-devkit/schematics';
-import { WorkspaceSchema } from '@angular-devkit/core/src/experimental/workspace';
+import {
+  WorkspaceProject,
+  WorkspaceSchema
+} from '@angular-devkit/core/src/experimental/workspace';
 import { ClassDeclaration, Node, Project } from 'ts-morph';
+import { resolve } from 'path';
 
 export const findAngularJSON = (tree: Tree): WorkspaceSchema => {
   const angularJson = tree.read('angular.json');
@@ -10,6 +14,35 @@ export const findAngularJSON = (tree: Tree): WorkspaceSchema => {
 
   const content = angularJson.toString();
   return JSON.parse(content) as WorkspaceSchema;
+};
+
+export const getProjectAST = (
+  workspace: WorkspaceProject,
+  projectName: string
+): Project => {
+  const tsConfigs: string | string[] =
+    workspace.architect?.build?.options?.tsConfig;
+  let configPath: string | undefined;
+  if (Array.isArray(tsConfigs)) {
+    configPath = tsConfigs?.find(
+      conf =>
+        conf.includes('tsconfig.lib') ||
+        conf.includes('tsconfig.json') ||
+        conf.includes('tsconfig.app.json')
+    );
+  } else if (typeof tsConfigs === 'string') {
+    configPath = tsConfigs;
+  } else {
+    throw new Error(
+      `The config is not specified in angular.json for ${projectName} project`
+    );
+  }
+
+  const absoluteConfigPath = resolve(process.cwd(), configPath as string);
+  return new Project({
+    tsConfigFilePath: absoluteConfigPath,
+    addFilesFromTsConfig: true
+  });
 };
 
 export const getRouterModuleClass = (project: Project): ClassDeclaration => {
